backend/requirements.txt
Flask
python-pptx
flask-cors
flask_sqlalchemy
google-generativeai>=0.8.3
transformers
torch
python-dotenv
redis==5.0.0
pymongo[srv]==4.7.0
psycopg2-binary
pgvector
googlesearch-python
bs4